Move virtual file types under Bicep.Testing.IO
