+++ src/Bicep.Core.UnitTests/Diagnostics/LinterRuleTests/NoUnusedImportsRuleTests.cs
@@ -5,6 +5,7 @@ using Bicep.Core.Analyzers.Linter.Rules;
 using Bicep.Core.UnitTests.Assertions;
 using Bicep.Core.UnitTests.Utils;
 using Bicep.Testing;
+using Bicep.Testing.IO;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
