+++ src/Bicep.Testing/Baselines/BaselineDirectory.cs
@@ -4,6 +4,7 @@
 using System.Collections.Immutable;
 using Bicep.Core.Extensions;
 using Bicep.Testing.Assertions;
+using Bicep.Testing.IO;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -15,7 +16,7 @@ public record BaselineDirectory(
     ImmutableDictionary<string, BaselineFile> Files,
     BaselineFile EntryFile)
 {
-    internal static BaselineDirectory Materialize(TestContext testContext, TestEmbeddedFile embeddedFile)
+    internal static BaselineDirectory Materialize(TestContext testContext, EmbeddedFile embeddedFile)
     {
         var outputDirectory = testContext.GetUniqueOutputPath();
         var baselines = embeddedFile.GetDirectoryFiles().ToImmutableDictionary(
@@ -44,7 +45,7 @@ public record BaselineDirectory(
             return baseline;
         }
 
-        var embeddedFile = new TestEmbeddedFile(
+        var embeddedFile = new EmbeddedFile(
             EntryFile.EmbeddedFile.Assembly,
             $"{StreamDirectoryPath}/{relativePath}");
