+++ src/Bicep.Testing/Baselines/BaselineFile.cs
@@ -2,10 +2,11 @@
 // Licensed under the MIT License.
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Bicep.Testing.IO;
 
 namespace Bicep.Testing.Baselines;
 
-public record BaselineFile(TestContext TestContext, TestEmbeddedFile EmbeddedFile, string OutputFilePath)
+public record BaselineFile(TestContext TestContext, EmbeddedFile EmbeddedFile, string OutputFilePath)
 {
     public string Read() => File.ReadAllText(OutputFilePath);
