+++ src/Bicep.Testing/IO/EmbeddedFile.cs
@@ -4,9 +4,9 @@
 using System.Reflection;
 using System.Text.RegularExpressions;
 
-namespace Bicep.Testing;
+namespace Bicep.Testing.IO;
 
-public record TestEmbeddedFile(Assembly Assembly, string StreamPath)
+public record EmbeddedFile(Assembly Assembly, string StreamPath)
 {
     private readonly Lazy<BinaryData> binaryDataLazy = new(() => BinaryData.FromStream(Assembly.GetManifestResourceStream(StreamPath)!));
     private readonly Lazy<string> contentsLazy = new(() => new StreamReader(Assembly.GetManifestResourceStream(StreamPath)!).ReadToEnd());
@@ -24,20 +24,20 @@ public record TestEmbeddedFile(Assembly Assembly, string StreamPath)
     public string GetPathRelativeToDirectory(string streamDirectoryPath)
         => StreamPath[streamDirectoryPath.Length..].TrimStart('/');
 
-    public IEnumerable<TestEmbeddedFile> GetDirectoryFiles()
+    public IEnumerable<EmbeddedFile> GetDirectoryFiles()
         => LoadAll(Assembly, streamPath => streamPath.StartsWith($"{StreamDirectoryPath}/", StringComparison.Ordinal));
 
-    public static IEnumerable<TestEmbeddedFile> LoadAll(Assembly assembly, string streamPathPrefix, Func<string, bool> shouldLoad)
+    public static IEnumerable<EmbeddedFile> LoadAll(Assembly assembly, string streamPathPrefix, Func<string, bool> shouldLoad)
     {
         var combinedPathPrefix = $"Files/{streamPathPrefix}/";
 
         return LoadAll(assembly, name => name.StartsWith(combinedPathPrefix, StringComparison.Ordinal) && shouldLoad(name));
     }
 
-    public static IEnumerable<TestEmbeddedFile> LoadAll(Assembly assembly, Regex regex)
+    public static IEnumerable<EmbeddedFile> LoadAll(Assembly assembly, Regex regex)
         => LoadAll(assembly, regex.IsMatch);
 
-    public static IEnumerable<TestEmbeddedFile> LoadAll(Assembly assembly, Func<string, bool> shouldLoad)
+    public static IEnumerable<EmbeddedFile> LoadAll(Assembly assembly, Func<string, bool> shouldLoad)
     {
         foreach (var streamName in assembly.GetManifestResourceNames().Where(shouldLoad))
         {
